Build AddResults component
